remove validator file and include inside handler
